Add more extension methods. Rename extensions-class.
